Revert varInt encoding for sparseSet
In memory varInt encoding wastes memory due to object and array overhead (8+4
bytes per byte[]). It makes no sense to use 13 to 18 bytes to store 4 bytes.

Memory overhead is still high due to the list of integers, autoboxing is very
expensive. We should replace it by an int array to keep the overhead low.