Faster addAll when other is in sparse mode.
addAll is much faster in sparse mode since only existing indexes have to be
updated. We can be smarter than creating a copy of other and converting it
to normal mode when other is in smart mode. We can traverse the other sparseSet
and only update relevant index.

This patch speed up scenario when lot of small cardinalities HLL++ are
aggregated into more coarse grained buckets.